scripts/requirements.txt
# requirements.txt
firebase-admin
python-dotenv
faker